backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cssselect2==0.6.0
defusedxml==0.7.1
Django==2.2.27
django-colorfield==0.7.1
django-extra-fields==3.0.2
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-base64==2.0
flake8==4.0.1
fonttools==4.34.4
gunicorn==20.0.4
html5lib==1.1
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
Pillow==9.2.0
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pycparser==2.21
pydyf==0.2.0
pyflakes==2.4.0
PyJWT==2.4.0
pyphen==0.12.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tinycss2==1.1.1
uritemplate==4.1.1
urllib3==1.26.10
weasyprint==56.0
webencodings==0.5.1
zopfli==0.2.1
